Deduplicate code by using an abstract base class
