+++ src/Graviton/DocumentBundle/DependencyInjection/Compiler/ExtRefMappingCompilerPass.php
@@ -9,7 +9,6 @@
 
 namespace Graviton\DocumentBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
@@ -17,27 +16,18 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
  * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
  * @link     http://swisscom.ch
  */
-class ExtRefMappingCompilerPass implements CompilerPassInterface
+class ExtRefMappingCompilerPass extends ExtRefCompilerPass
 {
     /**
-     * load services
+     * create mapping from services
      *
      * @param ContainerBuilder $container container builder
      *
      * @return void
      */
-    public function process(ContainerBuilder $container)
+    public function processServices(ContainerBuilder $container, $services)
     {
-        $map = [];
-        $gravitonServices = array_filter(
-            $container->getServiceIds(),
-            function ($id) {
-                return substr($id, 0, 8) == 'graviton' &&
-                    strpos($id, 'controller') !== false &&
-                    $id !== 'graviton.rest.controller';
-            }
-        );
-        foreach ($gravitonServices as $id) {
+        foreach ($services as $id) {
             list($ns, $bundle,, $doc) = explode('.', $id);
             $map[ucfirst($doc)] = implode('.', [$ns, $bundle, 'rest', $doc, 'get']);
         }
